Moving LocationCallback
