+++ src/Views/Columns/ImageColumn.php
@@ -7,15 +7,15 @@ use Rappasoft\LaravelLivewireTables\Exceptions\DataTableConfigurationException;
 use Rappasoft\LaravelLivewireTables\Views\Column;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Configuration\ImageColumnConfiguration;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Helpers\ImageColumnHelpers;
+use Rappasoft\LaravelLivewireTables\Views\Traits\HasLocationCallback;
 
 class ImageColumn extends Column
 {
     use ImageColumnConfiguration,
-        ImageColumnHelpers;
+        ImageColumnHelpers,
+        HasLocationCallback;
 
     protected string $view = 'livewire-tables::includes.columns.image';
-
-    protected mixed $locationCallback = null;
 
     public function __construct(string $title, ?string $from = null)
     {
