+++ src/Views/Columns/LinkColumn.php
@@ -7,17 +7,17 @@ use Rappasoft\LaravelLivewireTables\Exceptions\DataTableConfigurationException;
 use Rappasoft\LaravelLivewireTables\Views\Column;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Configuration\LinkColumnConfiguration;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Helpers\LinkColumnHelpers;
+use Rappasoft\LaravelLivewireTables\Views\Traits\HasLocationCallback;
 
 class LinkColumn extends Column
 {
     use LinkColumnConfiguration,
-        LinkColumnHelpers;
+        LinkColumnHelpers,
+        HasLocationCallback;
 
     protected string $view = 'livewire-tables::includes.columns.link';
 
     protected mixed $titleCallback = null;
-
-    protected mixed $locationCallback = null;
 
     public function __construct(string $title, ?string $from = null)
     {
