+++ src/Views/Traits/Configuration/ImageColumnConfiguration.php
@@ -4,11 +4,5 @@ namespace Rappasoft\LaravelLivewireTables\Views\Traits\Configuration;
 
 trait ImageColumnConfiguration
 {
-    // TODO: Test
-    public function location(callable $callback): self
-    {
-        $this->locationCallback = $callback;
 
-        return $this;
-    }
 }
