+++ src/Views/Traits/Helpers/ImageColumnHelpers.php
@@ -4,14 +4,5 @@ namespace Rappasoft\LaravelLivewireTables\Views\Traits\Helpers;
 
 trait ImageColumnHelpers
 {
-    // TODO: Test
-    public function getLocationCallback(): ?callable
-    {
-        return $this->locationCallback;
-    }
 
-    public function hasLocationCallback(): bool
-    {
-        return $this->locationCallback !== null;
-    }
 }
